refactor: split functionality into modules
